requirements.txt
streamlit
PyPDF2
python-dotenv
docx2txt
langchain_google_genai
google-generatveai